# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "lumpsum.in API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/learning/chapters": {
      "get": {
        "summary": "List learning chapters",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["BEGINNER", "INTERMEDIATE", "ADVANCED", "PROFESSIONAL"]
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["BASICS", "INVESTMENT", "DISTRIBUTOR", "RESEARCH"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of chapters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LearningChapter" }
                }
              }
            }
          }
        }
      }
    },
...
    "/learning/bookmarks/{id}": {
      "delete": {
        "summary": "Delete bookmark (requires JWT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Bookmark deleted" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LearningChapter": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "content": { "type": "string" },
          "level": {
            "type": "string",
            "enum": ["BEGINNER", "INTERMEDIATE", "ADVANCED", "PROFESSIONAL"]
          },
          "category": {
            "type": "string",
            "enum": ["BASICS", "INVESTMENT", "DISTRIBUTOR", "RESEARCH"]
          },
          "order": { "type": "integer" },
          "estimatedTime": { "type": "integer" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}